+++ ElvUI_LuckyoneUI/Core/Database.lua
@@ -1094,6 +1094,15 @@ function L1UI:Layout_Dragonflight(layout)
 		E.db.unitframe.units.party.roleIcon.yOffset = 0
 		E.db.unitframe.units.party.width = 94
 
+		-- Healing Party Pets [nonRetail only]
+		E.db.unitframe.units.party.petsGroup.anchorPoint = 'BOTTOM'
+		E.db.unitframe.units.party.petsGroup.enable = not E.Retail
+		E.db.unitframe.units.party.petsGroup.healPrediction.enable = true
+		E.db.unitframe.units.party.petsGroup.threatStyle = 'NONE'
+		E.db.unitframe.units.party.petsGroup.width = 94
+		E.db.unitframe.units.party.petsGroup.xOffset = 0
+		E.db.unitframe.units.party.petsGroup.yOffset = -1
+
 		-- Healing Raid1
 		E.db.unitframe.units.raid1.buffIndicator.enable = true
 		E.db.unitframe.units.raid1.height = 40
@@ -2035,7 +2044,6 @@ function L1UI:Layout_Shadowlands(layout)
 	E.db.movers.ThreatBarMover = 'BOTTOMRIGHT,ElvUIParent,BOTTOMRIGHT,-438,1'
 	E.db.movers.TooltipMover = 'BOTTOMRIGHT,ElvUIParent,BOTTOMRIGHT,0,135'
 	E.db.movers.TopCenterContainerMover = 'TOP,ElvUIParent,TOP,0,-57'
-	E.db.movers.TotemBarMover = 'BOTTOMLEFT,ElvUIParent,BOTTOMLEFT,428,1'
 	E.db.movers.VOICECHAT = 'TOPLEFT,ElvUIParent,TOPLEFT,1,-27'
 
 	if E.Retail then