+++ ElvUI_LuckyoneUI/Init.lua
@@ -44,6 +44,7 @@ local toons = {
 	["Luckyone-Everlook"] = true, -- Druid
 	["Luckypriest-Everlook"] = true, -- Priest
 	["Luckydk-Everlook"] = true, -- DK
+	["Luckyone-Giantstalker"] = true, -- Paladin
 }
 
 -- Constants
@@ -54,7 +55,7 @@ L1UI.DefaultTexture = 'Minimalist'
 L1UI.Logo = 'Interface\\AddOns\\ElvUI_LuckyoneUI\\Media\\Textures\\Clover.tga'
 L1UI.Me = toons[name]
 L1UI.Name = '|cff4beb2cLuckyoneUI|r'
-L1UI.RequiredVersion = 12.87
+L1UI.RequiredVersion = 12.89
 L1UI.Version = GetAddOnMetadata(addon, 'Version')
 
 function L1UI:Initialize()